Prevent class loading deadlock when a class in one group needs to load a class in another group.
The problem is caused by the default implementation of the ClassLoader.loadClass method which acquires a lock. The lock is the group instance itself, and the lock will also be acquired by the JVM before calling loadClass. This means that the lock acquisition order for the two groups is inconsistent. The fix is to call findLoadedClass and skip the second lock acquisition step.